requirements.txt
asgiref==3.6.0
Django==4.2
django-countries==7.5.1
geographiclib==2.0
geopy==2.3.0
psycopg2-binary==2.9.6
python-dotenv==0.21.0
sqlparse==0.4.3
typing_extensions==4.5.0
djangorestframework~=3.14.0
django-cors-headers==3.14.0
Pillow==9.5.0
django-filter==23.1
